backend/pyproject.toml
[project]
name = "yesand-backend"
version = "0.1.0"
description = "Collaborative image creation chatbot backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "openai>=1.50.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.hatch.build.targets.wheel]
packages = ["yesand"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
